+++ src/Math/Vector/4D/NRE_Vector4D.hpp
@@ -207,11 +207,11 @@
                         /**
                          * @return the vector's norm
                          */
-                        double norm() const;
+                        float norm() const;
                         /**
                          * @return the squared vector's norm
                          */
-                        double normSquared() const;
+                        float normSquared() const;
                         /**
                          * Normalize the vector
                          */
@@ -293,7 +293,7 @@
                          * @param u the vector
                          * @return  the scalar product
                          */
-                        double operator |=(Vector4D const& u) const;
+                        float operator |=(Vector4D const& u) const;
 
                     //## Arithmetic Operator ##//
                         /**
@@ -342,7 +342,7 @@
                          * @param u the vector
                          * @return  the scalar product
                          */
-                        double operator |(Vector4D const& u) const;
+                        float operator |(Vector4D const& u) const;
 
                     //## Comparison Operator ##//
                         /**
